Fix after restructuration.
Create page folder.
Add favicon.
