+++ src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
-import './App.css';
+import './style/App.css';
 
 import Navbar from './components/Navbar';
-import Home from './components/Home';
-import ChampionshipPage from './components/ChampionshipPage';
+import Home from './pages/Home';
+import ChampionshipPage from './pages/ChampionshipPage';
 
 const App: React.FC = (): JSX.Element => {
     // Set the capital here
